Move view/model code out of AppDelegate into a WindowController
